+++ Entidades/OtrasPrestaciones.cs
@@ -8,6 +8,6 @@ namespace WS_Avantius.Entidades
     public class OtrasPrestaciones
     {
         public List<OtraPrestacion> codigoOtraPrestacion {  get; set; }
-        public decimal totalOtrasPrestaciones { get; set; }
+        public double totalOtrasPrestaciones { get; set; }
     }
 }
